ev: rename ElectricVehicleSpecificationWithMatsimVehicle to ElectricVehicleSpecificationImpl
Now this is the default implementation.
